# balance/src/main/java/com/balance/balance/swagger/ConfigSwagger.yaml
openapi: "3.0.2"
info:
  title: Бонусная Система
  description: Просмотр баланса на карте
  version: 1.0.0
  contact:
    name: Георгий Панов
    email: panovgeorgij819@gmail.com
    url: https://georgij1.github.io/portfoilio_georgii/
servers:
  - url: https://localhost:8080
tags:
  - name: Баланс клиента
    description: Баланс на карте клиента
paths:
  /balance/getBalance/{CardId}:
    get:
      tags:
        - Баланс клиента
      summary: Баланс
      description: Получение баланса клиента
      operationId: getBalanceClient
      parameters:
        - name: CardId
          in: path
          description: CardId
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: ok
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BalanceTableDTO'
  /balance/createBalanceCustomerCard:
    post:
      tags:
        - Баланс клиента
      summary: Создание баланса клиента
      description: Позволяет создать баланс клиенту
      operationId: createBalanceCustomerCard
      requestBody:
        description: создание баланса клиента
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BalanceTableDTO'
        required: true
      responses:
        '201':
          description: ok

  /balance/updateBalanceCustomerCard:
    post:
      tags:
        - Баланс клиента
      summary: Обновление баланса клиента
      description: Позволяет обновить баланс клиенту
      operationId: updateBalanceCustomerCard
      requestBody:
        description: обновление баланса клиента
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateBalanceTableDto'
        required: true
      responses:
        '201':
          description: ok
  /balance/halfcheck:
    get:
      tags:
        - Баланс клиента
      summary: Статус сервера
      description: Запущен ли сервер
      operationId: half-check
      responses:
        '200':
          description: ok

components:
  schemas:
    BalanceTableDTO:
      required:
        - id
        - card_id
        - value
        - exp
      type: object
      properties:
        card_id:
          type: integer
        value:
          type: integer
        exp:
          type: string
    UpdateBalanceTableDto:
      required:
        - id
        - card_id
        - value
      type: object
      properties:
        id:
          type: string
          example: 0873ca6b-51fd-4708-b73d-43c023f0877d
        card_id:
          type: integer
          example: 1
        value:
          type: integer
          example: 100